+++ CHANGELOG.md
@@ -8,9 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Fixed
 - Switch to markdown links in the outputs table in the README.
+- Exclude samples if all the reads are removed during host depletion.
 ### Added
 - `igv` option to enable IGV in the EPI2ME Desktop Application.
 - `include_read_assignments` option to output a file with the taxonomy of each read.
+- `Reads` section in the report to track the number of reads after filtering, host depletion and unclassified.
 ### Changed
 - `keep_bam` is now only required to output BAM files.
 - `include_kraken2_assignments` has been replaced by `include_read_assignments`.
@@ -18,7 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - Kraken2 indexes to the ones released 2024-06-05
     - Taxonomy database to the one released 2024-09-01
 ### Removed
-- `split-prefix` parameter, as the workflow automatically enables this option for large reference genomes
+- `split-prefix` parameter, as the workflow automatically enables this option for large reference genomes.
 
 ## [v2.10.1]
 ### Fixed